+++ packages/frontapp/src/utils/routes.ts
@@ -9,7 +9,7 @@ export enum RoutesTypes {
   DONATES_FROM_USER = '/donates-from-user',
   CREATE_NEW_ITEM = '/create-new-item',
   DONATE_TO_LIBRARY = '/donate-to-library',
-  DONAT_PREVIEW = '/donat',
+  DONATE_PREVIEW = '/donate',
   BOOK_PREVIEW = 'item',
   CATEGORY = '/category',
 }
